# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: S.O.S Aid API
  version: 1.1.0
  description: >-
    Contract có hiệu lực của repo (Rule 6.1 – API First).

    Khác biệt so với bản tham chiếu trong Developer Kit
    (`docs/reference/openapi.kit-original.yaml`):
      * Mọi response bọc trong envelope `{success, data, requestId}` hoặc
        `{success, error, requestId}` theo Rule 6.2 / ADR-003.
      * `EmergencyCase` bổ sung `phase` – 6 pha hiển thị dẫn xuất từ `status`
        (ADR-001) – và `realtimeChannel`.
      * Bổ sung các endpoint đã triển khai: timeline, lịch sử vị trí, danh mục
        hướng dẫn, ứng viên điều phối, nhiệm vụ của kíp xe, audit, health.

    ĐÂY KHÔNG PHẢI API CHÍNH THỨC CỦA 115. Hệ thống chạy ở chế độ mô phỏng
    (`EMS_INTEGRATION_MODE=mock`) cho tới khi có thoả thuận/SLA chính thức.
servers:
  - url: http://localhost:3000/v1
    description: Local development
  - url: https://api.example.com/v1
    description: Placeholder – chưa có môi trường production
security:
  - bearerAuth: []
...
paths:
...
  /auth/otp/request:
    post:
      tags: [Auth]
      security: []
      summary: Yêu cầu mã OTP đăng nhập cho người dân
      description: >-
        Luôn trả 202 bất kể số điện thoại có tồn tại hay không, để endpoint này
        không trở thành công cụ dò tài khoản đã đăng ký.
        Có rate limit theo `RATE_LIMIT_OTP_PER_HOUR` (TC-026).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone]
              properties:
                phone: { type: string, example: '+84901234567' }
      responses:
        '202':
          description: Đã tiếp nhận yêu cầu gửi OTP
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          accepted: { type: boolean }
        '429': { $ref: '#/components/responses/RateLimited' }
  /auth/otp/verify:
    post:
      tags: [Auth]
      security: []
      summary: Xác thực OTP và cấp cặp token
      description: >-
        Tài khoản có vai trò nghiệp vụ (OPERATOR_115, CLINICIAN, AMBULANCE_CREW,
        FACILITY_USER, ADMIN, AUDITOR) BỊ TỪ CHỐI ở luồng này (SOS-004) và phải
        dùng cổng OIDC/MFA riêng.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, otp]
              properties:
                phone: { type: string }
                otp: { type: string, minLength: 6, maxLength: 6 }
      responses:
        '200':
          description: Đăng nhập thành công
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/LoginResult' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /auth/refresh:
    post:
      tags: [Auth]
      security: []
      summary: Đổi refresh token lấy cặp token mới (có rotation)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        '200':
          description: Cấp token mới; token cũ bị thu hồi ngay
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/LoginResult' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
...
  /emergency-cases/{caseId}/triage:
    post:
      tags: [Triage]
      summary: Gửi phiếu ghi nhận dấu hiệu quan sát được
      description: >-
        CHỈ lưu quan sát. Hệ thống KHÔNG chẩn đoán, KHÔNG chấm điểm mức độ,
        KHÔNG gợi ý xử trí (Rule 1.2). Không bắt buộc trả lời đủ mọi câu.
      parameters: [{ $ref: '#/components/parameters/CaseId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [questionnaireVersion, answers]
              properties:
                questionnaireVersion: { type: string, example: pilot-1 }
                answers:
                  type: array
                  minItems: 1
                  maxItems: 50
                  items:
                    type: object
                    required: [questionCode, value]
                    properties:
                      questionCode: { type: string, example: responsive }
                      value: { type: string, enum: [yes, no, unknown] }
      responses:
        '202': { $ref: '#/components/responses/GenericSuccess' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '409': { $ref: '#/components/responses/Conflict' }
    get:
      tags: [Triage]
      summary: Lịch sử phiếu quan sát của ca
      parameters: [{ $ref: '#/components/parameters/CaseId' }]
      responses:
        '200': { $ref: '#/components/responses/GenericSuccess' }

  # ==========================================================================
  # Video
  # ==========================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CaseId:
      in: path
      name: caseId
      required: true
      schema: { type: string, format: uuid }
  responses:
    GenericSuccess:
      description: Thành công
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiSuccess' }
    ValidationFailed:
      description: Dữ liệu không hợp lệ
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Unauthenticated:
      description: Chưa xác thực hoặc token không hợp lệ
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Forbidden:
      description: Không đủ quyền
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
...
    Conflict:
      description: Xung đột trạng thái/đồng thời
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    RateLimited:
      description: Vượt ngưỡng tần suất
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
...
  schemas:
...
    ApiSuccess:
      type: object
      required: [success, data, requestId]
      properties:
        success: { type: boolean, enum: [true] }
        data: {}
        requestId: { type: string }
    ApiError:
      type: object
      required: [success, error, requestId]
      properties:
        success: { type: boolean, enum: [false] }
        error:
          type: object
          required: [code, message]
          properties:
            code: { $ref: '#/components/schemas/ApiErrorCode' }
            message: { type: string, description: 'Thông điệp an toàn để hiển thị; không chứa stack trace' }
            details: { type: array, items: {} }
        requestId: { type: string }
    ApiErrorCode:
      type: string
      enum:
        - VALIDATION_FAILED
        - UNAUTHENTICATED
        - FORBIDDEN
        - NOT_FOUND
        - CONFLICT
        - CASE_INVALID_TRANSITION
        - CASE_ALREADY_ACCEPTED
        - IDEMPOTENCY_CONFLICT
        - GUIDANCE_NOT_APPROVED
        - CONSENT_REQUIRED
        - RATE_LIMITED
        - PROVIDER_UNAVAILABLE
        - INTERNAL_ERROR

    # --- Domain ------------------------------------------------------------
...
    LoginResult:
      type: object
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        expiresInSeconds: { type: integer }
        user:
          type: object
          properties:
            id: { type: string, format: uuid }
            fullName: { type: string, nullable: true }
            roles:
              type: array
              items:
                type: string
                enum: [CITIZEN, OPERATOR_115, CLINICIAN, AMBULANCE_CREW, FACILITY_USER, LOCAL_RESPONDER, ADMIN, AUDITOR]
...